Fix for issue 13.
git-svn-id: http://jmockit.googlecode.com/svn/trunk@509 4d38ed0c-b8d3-11de-a71d-8f2b2e058ad8

+++ main/test/mockit/ExpectationsWithArgMatchersTest.java
@@ -357,7 +357,7 @@ public final class ExpectationsWithArgMatchersTest
       new Expectations()
       {
          {
-            mock.setValue((Integer) with(IsEqual.equalTo(3)));
+            mock.setValue(with(IsEqual.equalTo(3)));
          }
       };
